pyproject.toml
[project]
name = "openbb-pydantic-ai"
version = "0.1.7"
description = "Pydantic AI adapter for OpenBB SDK. Enables connettion any pydantic-ai agent to the OpenBB Workspace."
authors = [
    { name = "Magnus Samuelsen" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openbb-ai>=1.8.5",
    "pydantic-ai-slim[ui]>=1.47.0",
]

[project.optional-dependencies]
pdf = [
    "docling>=2.72.0",
    "httpx>=0.28.1",
    "torch",
    "torchvision",
]
pdf-cu128 = [
    "docling>=2.72.0",
    "httpx>=0.28.1",
    "torch",
    "torchvision",
]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "dirty-equals>=0.10.0",
    "inline-snapshot>=0.31.0",
    "pre-commit>=4.4.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.4",
    "ty>=0.0.2",
]

[tool.uv]
conflicts = [
    [
        { extra = "pdf" },
        { extra = "pdf-cu128" },
    ],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "pdf" },
    { index = "pytorch-cu128", extra = "pdf-cu128" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "pdf" },
    { index = "pytorch-cu128", extra = "pdf-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "openbb_pydantic_ai"
module-root = ""

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "B", # flake8-bugbear
    "I", # isort
    "S", # security / flake8-bandit
]
ignore = [
    "S101", # ignore assert statements being used (for pytest)
]

[tool.pytest.ini_options]
pythonpath = [
    ".",
]
markers = [
    "regression_contract: high-risk adapter/event-stream behavior contract tests",
]
asyncio_mode = "auto"

[tool.ty.src]
exclude = ["tests/smoke/test_distribution.py"]
